backend/requirements.txt
fastapi>=0.135.1
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.48
pymysql>=1.1.1
cryptography>=44.0.0
python-multipart>=0.0.20
python-dotenv>=1.0.1
pillow>=11.0.0
imagehash>=4.3.1
numpy>=2.2.0
httpx>=0.28.0
anthropic>=0.40.0
openai>=1.57.0
google-generativeai>=0.8.3
aiofiles>=24.1.0
openpyxl>=3.1.5
alembic>=1.14.0
pydantic>=2.12.5
pydantic-settings>=2.7.0
requests>=2.32.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# For video frame extraction:
# opencv-python-headless>=4.9.0.0
# imageio>=2.34.0
